pyproject.toml
[project]
name = "budget"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "celery",
    "django",
    "django-celery-beat",
    "django-celery-results",
    "django-extensions>=4.1",
    "djangorestframework",
    "drf-spectacular",
    "matplotlib",
    "pandas",
    "pre-commit>=4.5.1",
    "psycopg[binary]",
    "pillow",
    "redis",
    "ruff>=0.14.10",
    "gunicorn",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=1.2.1",
]



[dependency-groups]
dev = [
    "ruff",
    "python-dotenv",
    "pre-commit",
]
prod = [
    "gunicorn",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "F403"]
